+++ tensorflow/contrib/persona/kernels/protein-cluster/ssw/ssw.h
@@ -126,7 +126,7 @@ void init_destroy (s_profile* p);
 			0-based coordinate.
 */
 s_align* ssw_align (const s_profile* prof,
-					const int8_t* ref,
+					const int32_t* ref,
 					int32_t refLen,
 					const uint8_t weight_gapO,
 					const uint8_t weight_gapE,
